@@ -1,5 +1,6 @@
 // mautrix-slack - A Matrix-Slack puppeting bridge.
 // Copyright (C) 2022 Max Sandholm
+// Copyright (C) 2024 Tulir Asokan
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as published by
@@ -14,38 +15,41 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package main
+package msgconv
 
 import (
+	"context"
 	"fmt"
-	"io/ioutil"
+	"html"
+	"io"
 	"net/http"
 	"path"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/slack-go/slack"
-	"github.com/yuin/goldmark"
+	"github.com/rs/zerolog"
 	"maunium.net/go/mautrix/event"
 	"maunium.net/go/mautrix/format"
 
-	"go.mau.fi/mautrix-slack/database"
+	"github.com/slack-go/slack"
+	"go.mau.fi/mautrix-slack/msgconv/mrkdwn"
 )
 
-func (portal *Portal) renderImageBlock(block slack.ImageBlock) (*event.MessageEventContent, error) {
+func (mc *MessageConverter) renderImageBlock(ctx context.Context, block slack.ImageBlock) (*ConvertedMessagePart, error) {
+	log := zerolog.Ctx(ctx)
 	client := http.Client{}
 	resp, err := client.Get(block.ImageURL)
 	if err != nil {
-		portal.log.Errorfln("Error fetching image: %v", err)
+		log.Err(err).Msg("Failed to fetch image block")
 		return nil, err
 	} else if resp.StatusCode != 200 {
-		portal.log.Errorfln("HTTP error %d fetching image", resp.StatusCode)
+		log.Error().Int("status_code", resp.StatusCode).Msg("Unexpected status code fetching image block")
 		return nil, fmt.Errorf(resp.Status)
 	}
-	bytes, err := ioutil.ReadAll(resp.Body)
+	bytes, err := io.ReadAll(resp.Body)
 	if err != nil {
-		portal.log.Errorfln("Error fetching image: %v", err)
+		log.Err(err).Msg("Failed to read image block get response")
 		return nil, err
 	}
 	filename := path.Base(resp.Request.URL.Path)
@@ -58,93 +62,86 @@ func (portal *Portal) renderImageBlock(block slack.ImageBlock) (*event.MessageEv
 			Size:     len(bytes),
 		},
 	}
-	err = portal.uploadMedia(portal.MainIntent(), bytes, &content)
+	err = mc.uploadMedia(ctx, bytes, &content)
 	if err != nil {
-		portal.log.Errorfln("Error uploading media: %v", err)
+		log.Err(err).Msg("Failed to reupload image block media")
 		return nil, err
 	}
-	return &content, nil
+	return &ConvertedMessagePart{
+		Type:    event.EventMessage,
+		Content: &content,
+	}, nil
 }
 
-func (portal *Portal) mrkdwnToMatrixHtml(mrkdwn string) string {
-	mrkdwn = replaceShortcodesWithEmojis(mrkdwn)
-
-	mrkdwn = escapeFixer.ReplaceAllStringFunc(mrkdwn, func(s string) string {
-		return s[:2] + `\` + s[2:]
-	})
-
-	mdRenderer := goldmark.New(
-		format.Extensions, format.HTMLOptions,
-		goldmark.WithExtensions(&SlackTag{portal}),
-	)
-
-	var buf strings.Builder
-	mdRenderer.Convert([]byte(mrkdwn), &buf)
-
-	return format.UnwrapSingleParagraph(buf.String())
+func (mc *MessageConverter) mrkdwnToMatrixHtml(ctx context.Context, inputMrkdwn string) string {
+	output, _ := mc.SlackMrkdwnParser.Parse(ctx, inputMrkdwn)
+	return output
 }
 
-func (portal *Portal) renderSlackTextBlock(block slack.TextBlockObject) string {
+func (mc *MessageConverter) renderSlackTextBlock(ctx context.Context, block slack.TextBlockObject) string {
 	if block.Type == slack.PlainTextType {
 		return event.TextToHTML(block.Text)
 	} else if block.Type == slack.MarkdownType {
-		return portal.mrkdwnToMatrixHtml(block.Text)
+		return mc.mrkdwnToMatrixHtml(ctx, block.Text)
 	} else {
 		return ""
 	}
 }
 
-func (portal *Portal) renderRichTextSectionElements(elements []slack.RichTextSectionElement, userTeam *database.UserTeam) string {
+func openingTags(out io.StringWriter, style *slack.RichTextSectionTextStyle) {
+	if style == nil {
+		return
+	}
+	if style.Bold {
+		_, _ = out.WriteString("<strong>")
+	}
+	if style.Italic {
+		_, _ = out.WriteString("<em>")
+	}
+	if style.Strike {
+		_, _ = out.WriteString("<del>")
+	}
+	if style.Code {
+		_, _ = out.WriteString("<code>")
+	}
+}
+
+func closingTags(out io.StringWriter, style *slack.RichTextSectionTextStyle) {
+	if style == nil {
+		return
+	}
+	if style.Code {
+		_, _ = out.WriteString("</code>")
+	}
+	if style.Strike {
+		_, _ = out.WriteString("</del>")
+	}
+	if style.Italic {
+		_, _ = out.WriteString("</em>")
+	}
+	if style.Bold {
+		_, _ = out.WriteString("</strong>")
+	}
+}
+
+func (mc *MessageConverter) renderRichTextSectionElements(ctx context.Context, elements []slack.RichTextSectionElement) string {
 	var htmlText strings.Builder
 	for _, element := range elements {
 		switch e := element.(type) {
 		case *slack.RichTextSectionTextElement:
-			if e.Style != nil {
-				if e.Style.Bold {
-					htmlText.WriteString("<b>")
-				}
-				if e.Style.Italic {
-					htmlText.WriteString("<i>")
-				}
-				if e.Style.Strike {
-					htmlText.WriteString("<del>")
-				}
-				if e.Style.Code {
-					htmlText.WriteString("<code>")
-				}
-			}
+			openingTags(&htmlText, e.Style)
 			htmlText.WriteString(event.TextToHTML(e.Text))
-			if e.Style != nil {
-				if e.Style.Code {
-					htmlText.WriteString("</code>")
-				}
-				if e.Style.Strike {
-					htmlText.WriteString("</del>")
-				}
-				if e.Style.Italic {
-					htmlText.WriteString("</i>")
-				}
-				if e.Style.Bold {
-					htmlText.WriteString("</b>")
-				}
-			}
+			closingTags(&htmlText, e.Style)
 		case *slack.RichTextSectionUserElement:
-			puppet := portal.bridge.GetPuppetByID(portal.Key.TeamID, e.UserID)
-			if puppet != nil && puppet.GetCustomOrGhostMXID() != "" {
-				htmlText.WriteString(fmt.Sprintf(`<a href="https://matrix.to/#/%s">%s</a>`, puppet.GetCustomOrGhostMXID(), puppet.Name))
-			} else { // TODO: register puppet and get info if not exist
-				htmlText.WriteString(fmt.Sprintf("@%s", e.UserID))
-			}
+			mxid, name := mc.GetMentionedUserInfo(ctx, e.UserID)
+			openingTags(&htmlText, e.Style)
+			mrkdwn.UserMentionToHTML(&htmlText, e.UserID, mxid, name)
+			closingTags(&htmlText, e.Style)
 		case *slack.RichTextSectionChannelElement:
-			p := portal.bridge.DB.Portal.GetByID(database.PortalKey{
-				TeamID:    portal.Key.TeamID,
-				ChannelID: e.ChannelID,
-			})
-			if p != nil && p.MXID != "" {
-				htmlText.WriteString(fmt.Sprintf(`<a href="https://matrix.to/#/%s?via=%s">%s</a>`, p.MXID, portal.bridge.AS.HomeserverDomain, p.Name))
-			} else { // TODO: get portal info if not exist
-				htmlText.WriteString(fmt.Sprintf("#%s", e.ChannelID))
-			}
+			mxid, alias, name := mc.GetMentionedRoomInfo(ctx, e.ChannelID)
+			openingTags(&htmlText, e.Style)
+			mrkdwn.RoomMentionToHTML(&htmlText, e.ChannelID, mxid, alias, name, mc.ServerName)
+			closingTags(&htmlText, e.Style)
 		case *slack.RichTextSectionLinkElement:
 			var linkText string
 			if e.Text != "" {
@@ -152,48 +149,54 @@ func (portal *Portal) renderRichTextSectionElements(elements []slack.RichTextSec
 			} else {
 				linkText = e.URL
 			}
-			htmlText.WriteString(fmt.Sprintf(`<a href="%s">%s</a>`, e.URL, event.TextToHTML(linkText)))
+			openingTags(&htmlText, e.Style)
+			_, _ = fmt.Fprintf(&htmlText, `<a href="%s">%s</a>`, html.EscapeString(e.URL), event.TextToHTML(linkText))
+			closingTags(&htmlText, e.Style)
 		case *slack.RichTextSectionBroadcastElement:
 			htmlText.WriteString("@room")
 		case *slack.RichTextSectionEmojiElement:
+			openingTags(&htmlText, e.Style)
 			if e.Unicode != "" {
 				codepoints := strings.Split(e.Unicode, "-")
 				for _, codepoint := range codepoints {
 					codepointInt, _ := strconv.ParseInt(codepoint, 16, 32)
-					unquoted := string(rune(codepointInt))
-					htmlText.WriteString(unquoted)
+					htmlText.WriteRune(rune(codepointInt))
 				}
 			} else {
-				emoji := portal.bridge.GetEmoji(e.Name, userTeam)
+				emoji := mc.GetEmoji(ctx, e.Name)
 				if strings.HasPrefix(emoji, "mxc://") {
-					htmlText.WriteString(fmt.Sprintf(`<img data-mx-emoticon src="%[1]s" alt="%[2]s" title="%[2]s" height="32"/>`, emoji, e.Name))
+					htmlText.WriteString(fmt.Sprintf(`<img data-mx-emoticon src="%[1]s" alt=":%[2]s:" title=":%[2]s:" height="32"/>`, emoji, e.Name))
 				} else if emoji != e.Name {
 					htmlText.WriteString(emoji)
 				} else {
 					htmlText.WriteString(fmt.Sprintf(":%s:", e.Name))
 				}
 			}
+			closingTags(&htmlText, e.Style)
 		case *slack.RichTextSectionColorElement:
 			htmlText.WriteString(e.Value)
 		case *slack.RichTextSectionDateElement:
 			htmlText.WriteString(e.Timestamp.String())
 		default:
-			portal.log.Warnfln("Slack rich text section contained unknown element %s", e.RichTextSectionElementType())
+			zerolog.Ctx(ctx).Debug().
+				Type("section_type", e).
+				Str("section_type_name", string(e.RichTextSectionElementType())).
+				Msg("Unsupported Slack rich text section")
 		}
 	}
 	return htmlText.String()
 }
 
-func (portal *Portal) renderSlackBlock(block slack.Block, userTeam *database.UserTeam) (string, bool) {
+func (mc *MessageConverter) renderSlackBlock(ctx context.Context, block slack.Block) (string, bool) {
 	switch b := block.(type) {
 	case *slack.HeaderBlock:
-		return fmt.Sprintf("<h1>%s</h1>", portal.renderSlackTextBlock(*b.Text)), false
+		return fmt.Sprintf("<h1>%s</h1>", mc.renderSlackTextBlock(ctx, *b.Text)), false
 	case *slack.DividerBlock:
 		return "<hr>", false
 	case *slack.SectionBlock:
 		var htmlParts []string
 		if b.Text != nil {
-			htmlParts = append(htmlParts, portal.renderSlackTextBlock(*b.Text))
+			htmlParts = append(htmlParts, mc.renderSlackTextBlock(ctx, *b.Text))
 		}
 		if len(b.Fields) > 0 {
 			var fieldTable strings.Builder
@@ -202,7 +205,7 @@ func (portal *Portal) renderSlackBlock(block slack.Block, userTeam *database.Use
 				if i%2 == 0 {
 					fieldTable.WriteString("<tr>")
 				}
-				fieldTable.WriteString(fmt.Sprintf("<td>%s</td>", portal.mrkdwnToMatrixHtml(field.Text)))
+				fieldTable.WriteString(fmt.Sprintf("<td>%s</td>", mc.mrkdwnToMatrixHtml(ctx, field.Text)))
 				if i%2 != 0 || i == len(b.Fields)-1 {
 					fieldTable.WriteString("</tr>")
 				}
@@ -214,7 +217,7 @@ func (portal *Portal) renderSlackBlock(block slack.Block, userTeam *database.Use
 	case *slack.RichTextBlock:
 		var htmlText strings.Builder
 		for _, element := range b.Elements {
-			htmlText.WriteString(portal.renderSlackRichTextElement(len(b.Elements), element, userTeam))
+			htmlText.WriteString(mc.renderSlackRichTextElement(ctx, len(b.Elements), element))
 		}
 		return format.UnwrapSingleParagraph(htmlText.String()), false
 	case *slack.ContextBlock:
@@ -222,21 +225,27 @@ func (portal *Portal) renderSlackBlock(block slack.Block, userTeam *database.Use
 		var unsupported bool = false
 		for _, element := range b.ContextElements.Elements {
 			if mrkdwnElem, ok := element.(*slack.TextBlockObject); ok {
-				htmlText.WriteString(fmt.Sprintf("<sup>%s</sup>", portal.mrkdwnToMatrixHtml(mrkdwnElem.Text)))
+				htmlText.WriteString(fmt.Sprintf("<sup>%s</sup>", mc.mrkdwnToMatrixHtml(ctx, mrkdwnElem.Text)))
 			} else {
-				portal.log.Debugfln("Unsupported Slack block element: %s", element.MixedElementType())
+				zerolog.Ctx(ctx).Debug().
+					Type("element_type", element).
+					Type("element_type_name", element.MixedElementType()).
+					Msg("Unsupported Slack block element")
 				htmlText.WriteString("<i>Slack message contains unsupported elements.</i>")
 				unsupported = true
 			}
 		}
 		return htmlText.String(), unsupported
 	default:
-		portal.log.Debugfln("Unsupported Slack block: %s", b.BlockType())
+		zerolog.Ctx(ctx).Debug().
+			Type("block_type", b).
+			Type("block_type_name", b.BlockType()).
+			Msg("Unsupported Slack block")
 		return "<i>Slack message contains unsupported elements.</i>", true
 	}
 }
 
-func (portal *Portal) renderSlackRichTextElement(numElements int, element slack.RichTextElement, userTeam *database.UserTeam) string {
+func (mc *MessageConverter) renderSlackRichTextElement(ctx context.Context, numElements int, element slack.RichTextElement) string {
 	switch e := element.(type) {
 	case *slack.RichTextSection:
 		var htmlTag string
@@ -251,7 +260,7 @@ func (portal *Portal) renderSlackRichTextElement(numElements int, element slack.
 			htmlTag = "<p>"
 			htmlCloseTag = "</p>"
 		}
-		return fmt.Sprintf("%s%s%s", htmlTag, portal.renderRichTextSectionElements(e.Elements, userTeam), htmlCloseTag)
+		return fmt.Sprintf("%s%s%s", htmlTag, mc.renderRichTextSectionElements(ctx, e.Elements), htmlCloseTag)
 	case *slack.RichTextList:
 		var htmlText strings.Builder
 		var htmlTag string
@@ -265,27 +274,27 @@ func (portal *Portal) renderSlackRichTextElement(numElements int, element slack.
 		}
 		htmlText.WriteString(htmlTag)
 		for _, e := range e.Elements {
-			htmlText.WriteString(fmt.Sprintf("<li>%s</li>", portal.renderSlackRichTextElement(1, &e, userTeam)))
+			htmlText.WriteString(fmt.Sprintf("<li>%s</li>", mc.renderSlackRichTextElement(ctx, 1, &e)))
 		}
 		htmlText.WriteString(htmlCloseTag)
 		return htmlText.String()
 	default:
-		portal.log.Debugfln("Unsupported Slack section: %T", e)
+		zerolog.Ctx(ctx).Debug().Type("element_type", e).Msg("Unsupported Slack rich text element")
 		return fmt.Sprintf("<i>Unsupported section %s in Slack text.</i>", e.RichTextElementType())
 	}
 }
 
-func (portal *Portal) blocksToHtml(blocks slack.Blocks, alwaysWrap bool, userTeam *database.UserTeam) string {
+func (mc *MessageConverter) blocksToHTML(ctx context.Context, blocks slack.Blocks, alwaysWrap bool) string {
 	var htmlText strings.Builder
 
 	if len(blocks.BlockSet) == 1 && !alwaysWrap {
 		// don't wrap in <p> tag if there's only one block
-		text, _ := portal.renderSlackBlock(blocks.BlockSet[0], userTeam)
+		text, _ := mc.renderSlackBlock(ctx, blocks.BlockSet[0])
 		htmlText.WriteString(text)
 	} else {
 		var lastBlockWasUnsupported bool = false
 		for _, block := range blocks.BlockSet {
-			text, unsupported := portal.renderSlackBlock(block, userTeam)
+			text, unsupported := mc.renderSlackBlock(ctx, block)
 			if !(unsupported && lastBlockWasUnsupported) {
 				htmlText.WriteString(fmt.Sprintf("<p>%s</p>", text))
 			}
@@ -296,19 +305,34 @@ func (portal *Portal) blocksToHtml(blocks slack.Blocks, alwaysWrap bool, userTea
 	return htmlText.String()
 }
 
-func (portal *Portal) SlackBlocksToMatrix(blocks slack.Blocks, attachments []slack.Attachment, userTeam *database.UserTeam) (*event.MessageEventContent, error) {
+func (mc *MessageConverter) trySlackBlocksToMatrix(ctx context.Context, blocks slack.Blocks, attachments []slack.Attachment) *ConvertedMessagePart {
+	converted, err := mc.slackBlocksToMatrix(ctx, blocks, attachments)
+	if err != nil {
+		zerolog.Ctx(ctx).Err(err).Msg("Failed to render Slack blocks")
+		return &ConvertedMessagePart{
+			Type: event.EventMessage,
+			Content: &event.MessageEventContent{
+				MsgType: event.MsgNotice,
+				Body:    "Failed to convert Slack message blocks",
+			},
+		}
+	}
+	return converted
+}
+
+func (mc *MessageConverter) slackBlocksToMatrix(ctx context.Context, blocks slack.Blocks, attachments []slack.Attachment) (*ConvertedMessagePart, error) {
 
 	// Special case for bots like the Giphy bot which send images in a specific format
 	if len(blocks.BlockSet) == 2 &&
 		blocks.BlockSet[0].BlockType() == slack.MBTImage &&
 		blocks.BlockSet[1].BlockType() == slack.MBTContext {
 		imageBlock := blocks.BlockSet[0].(*slack.ImageBlock)
-		return portal.renderImageBlock(*imageBlock)
+		return mc.renderImageBlock(ctx, *imageBlock)
 	}
 
 	var htmlText strings.Builder
 
-	htmlText.WriteString(portal.blocksToHtml(blocks, false, userTeam))
+	htmlText.WriteString(mc.blocksToHTML(ctx, blocks, false))
 
 	if len(attachments) > 0 && htmlText.String() != "" {
 		htmlText.WriteString("<br>")
@@ -317,18 +341,18 @@ func (portal *Portal) SlackBlocksToMatrix(blocks slack.Blocks, attachments []sla
 	for _, attachment := range attachments {
 		if attachment.IsMsgUnfurl {
 			for _, message_block := range attachment.MessageBlocks {
-				renderedAttachment := portal.blocksToHtml(message_block.Message.Blocks, true, userTeam)
+				renderedAttachment := mc.blocksToHTML(ctx, message_block.Message.Blocks, true)
 				htmlText.WriteString(fmt.Sprintf("<blockquote><b>%s</b><br>%s<a href=\"%s\"><i>%s</i></a><br></blockquote>",
 					attachment.AuthorName, renderedAttachment, attachment.FromURL, attachment.Footer))
 			}
 		} else if len(attachment.Blocks.BlockSet) > 0 {
 			for _, message_block := range attachment.Blocks.BlockSet {
-				renderedAttachment, _ := portal.renderSlackBlock(message_block, userTeam)
+				renderedAttachment, _ := mc.renderSlackBlock(ctx, message_block)
 				htmlText.WriteString(fmt.Sprintf("<blockquote>%s</blockquote>", renderedAttachment))
 			}
 		} else {
 			if len(attachment.Pretext) > 0 {
-				htmlText.WriteString(fmt.Sprintf("<p>%s</p>", portal.mrkdwnToMatrixHtml(attachment.Pretext)))
+				htmlText.WriteString(fmt.Sprintf("<p>%s</p>", mc.mrkdwnToMatrixHtml(ctx, attachment.Pretext)))
 			}
 			var attachParts []string
 			if len(attachment.AuthorName) > 0 {
@@ -342,15 +366,15 @@ func (portal *Portal) SlackBlocksToMatrix(blocks slack.Blocks, attachments []sla
 			if len(attachment.Title) > 0 {
 				if len(attachment.TitleLink) > 0 {
 					attachParts = append(attachParts, fmt.Sprintf("<b><a href=\"%s\">%s</a></b>",
-						attachment.TitleLink, portal.mrkdwnToMatrixHtml(attachment.Title)))
+						attachment.TitleLink, mc.mrkdwnToMatrixHtml(ctx, attachment.Title)))
 				} else {
-					attachParts = append(attachParts, fmt.Sprintf("<b>%s</b>", portal.mrkdwnToMatrixHtml(attachment.Title)))
+					attachParts = append(attachParts, fmt.Sprintf("<b>%s</b>", mc.mrkdwnToMatrixHtml(ctx, attachment.Title)))
 				}
 			}
 			if len(attachment.Text) > 0 {
-				attachParts = append(attachParts, portal.mrkdwnToMatrixHtml(attachment.Text))
+				attachParts = append(attachParts, mc.mrkdwnToMatrixHtml(ctx, attachment.Text))
 			} else if len(attachment.Fallback) > 0 {
-				attachParts = append(attachParts, portal.mrkdwnToMatrixHtml(attachment.Fallback))
+				attachParts = append(attachParts, mc.mrkdwnToMatrixHtml(ctx, attachment.Fallback))
 			}
 			htmlText.WriteString(fmt.Sprintf("<blockquote>%s", strings.Join(attachParts, "<br>")))
 			if len(attachment.Fields) > 0 {
@@ -361,7 +385,7 @@ func (portal *Portal) SlackBlocksToMatrix(blocks slack.Blocks, attachments []sla
 						fieldBody += "<tr>"
 					}
 					fieldBody += fmt.Sprintf("<td><strong>%s</strong><br>%s</td>",
-						field.Title, portal.mrkdwnToMatrixHtml(field.Value))
+						field.Title, mc.mrkdwnToMatrixHtml(ctx, field.Value))
 					short = !short && field.Short
 					if !short {
 						fieldBody += "</tr>"
@@ -373,7 +397,7 @@ func (portal *Portal) SlackBlocksToMatrix(blocks slack.Blocks, attachments []sla
 			}
 			var footerParts []string
 			if len(attachment.Footer) > 0 {
-				footerParts = append(footerParts, portal.mrkdwnToMatrixHtml(attachment.Footer))
+				footerParts = append(footerParts, mc.mrkdwnToMatrixHtml(ctx, attachment.Footer))
 			}
 			if len(attachment.Ts) > 0 {
 				ts, _ := attachment.Ts.Int64()
@@ -388,5 +412,8 @@ func (portal *Portal) SlackBlocksToMatrix(blocks slack.Blocks, attachments []sla
 	}
 
 	content := format.HTMLToContent(htmlText.String())
-	return &content, nil
+	return &ConvertedMessagePart{
+		Type:    event.EventMessage,
+		Content: &content,
+	}, nil
 }
