@@ -1,4 +1,4 @@
-// mautrix-whatsapp - A Matrix-WhatsApp puppeting bridge.
+// mautrix-slack - A Matrix-Slack puppeting bridge.
 // Copyright (C) 2021, 2022 Tulir Asokan, Sumner Evans, Max Sandholm
 //
 // This program is free software: you can redistribute it and/or modify
@@ -17,71 +17,35 @@
 package database
 
 import (
-	"database/sql"
-	"errors"
+	"context"
 
-	log "maunium.net/go/maulogger/v2"
-
-	"maunium.net/go/mautrix/util/dbutil"
+	"go.mau.fi/util/dbutil"
 )
 
 type BackfillQuery struct {
-	db  *Database
-	log log.Logger
+	*dbutil.QueryHelper[*BackfillState]
 }
 
-func (bq *BackfillQuery) NewBackfillState(portalKey *PortalKey) *BackfillState {
-	return &BackfillState{
-		db:     bq.db,
-		log:    bq.log,
-		Portal: portalKey,
-	}
+func newBackfill(qh *dbutil.QueryHelper[*BackfillState]) *BackfillState {
+	return &BackfillState{qh: qh}
 }
 
 const (
 	getBackfillState = `
 		SELECT team_id, channel_id, dispatched, backfill_complete, message_count, immediate_complete
 		FROM backfill_state
-		WHERE team_id=$1
-			AND channel_id=$2
+		WHERE team_id=$1 AND channel_id=$2
 	`
-
 	getNextUnfinishedBackfillState = `
 		SELECT team_id, channel_id, dispatched, backfill_complete, message_count, immediate_complete
 		FROM backfill_state
 		WHERE dispatched IS FALSE
 		AND backfill_complete IS FALSE
-		AND immediate_complete=$1
-		ORDER BY message_count ASC
+		ORDER BY CASE WHEN immediate_complete THEN 0 ELSE 1 END, message_count
 		LIMIT 1
 	`
-)
-
-type BackfillState struct {
-	db  *Database
-	log log.Logger
-
-	// Fields
-	Portal            *PortalKey
-	Dispatched        bool
-	BackfillComplete  bool
-	MessageCount      int
-	ImmediateComplete bool
-}
-
-func (b *BackfillState) Scan(row dbutil.Scannable) *BackfillState {
-	err := row.Scan(&b.Portal.TeamID, &b.Portal.ChannelID, &b.Dispatched, &b.BackfillComplete, &b.MessageCount, &b.ImmediateComplete)
-	if err != nil {
-		if !errors.Is(err, sql.ErrNoRows) {
-			b.log.Errorln("Database scan failed:", err)
-		}
-		return nil
-	}
-	return b
-}
-
-func (b *BackfillState) Upsert() {
-	_, err := b.db.Exec(`
+	undispatchAllBackfillsQuery = `UPDATE backfill_state SET dispatched=false`
+	upsertBackfillStateQuery    = `
 		INSERT INTO backfill_state
 			(team_id, channel_id, dispatched, backfill_complete, message_count, immediate_complete)
 		VALUES ($1, $2, $3, $4, $5, $6)
@@ -90,59 +54,49 @@ func (b *BackfillState) Upsert() {
 			dispatched=EXCLUDED.dispatched,
 			backfill_complete=EXCLUDED.backfill_complete,
 			message_count=EXCLUDED.message_count,
-			immediate_complete=EXCLUDED.immediate_complete`,
-		b.Portal.TeamID, b.Portal.ChannelID, b.Dispatched, b.BackfillComplete, b.MessageCount, b.ImmediateComplete)
-	if err != nil {
-		b.log.Warnfln("Failed to insert backfill state for %s: %v", b.Portal, err)
-	}
-}
+			immediate_complete=EXCLUDED.immediate_complete
+	`
+)
 
-func (b *BackfillState) SetDispatched(d bool) {
-	b.Dispatched = d
-	b.Upsert()
-}
-
-// Undispatch backfills so they can be retried in case the bridge crashed/was stopped during backfill
+// UndispatchAll undispatches backfills so they can be retried in case the bridge crashed/was stopped during backfill
 // Sent messages are tracked in the message and portal tables so this shouldn't lead to duplicate backfills
-func (b *BackfillQuery) UndispatchAll() {
-	_, err := b.db.Exec(`UPDATE backfill_state SET dispatched=FALSE`)
-	if err != nil {
-		b.log.Warnfln("Failed to undispatch all currently dispatched backfills")
-	}
+func (bq *BackfillQuery) UndispatchAll(ctx context.Context) error {
+	return bq.Exec(ctx, undispatchAllBackfillsQuery)
 }
 
-func (bq *BackfillQuery) GetBackfillState(portalKey *PortalKey) (backfillState *BackfillState) {
-	rows, err := bq.db.Query(getBackfillState, portalKey.TeamID, portalKey.ChannelID)
-	if err != nil || rows == nil {
-		bq.log.Error(err)
-		return
-	}
-	defer rows.Close()
-	if rows.Next() {
-		backfillState = bq.NewBackfillState(portalKey).Scan(rows)
-	}
-	return
+func (bq *BackfillQuery) GetBackfillState(ctx context.Context, portalKey PortalKey) (*BackfillState, error) {
+	return bq.QueryOne(ctx, getBackfillState, portalKey.TeamID, portalKey.ChannelID)
 }
 
-func (bq *BackfillQuery) GetNextUnfinishedBackfillState() (backfillState *BackfillState) {
-	rows, err := bq.db.Query(getNextUnfinishedBackfillState, false)
-	if err != nil || rows == nil {
-		bq.log.Error(err)
-		return
-	}
-	defer rows.Close()
-	if rows.Next() {
-		backfillState = bq.NewBackfillState(&PortalKey{}).Scan(rows)
-	} else {
-		rows, err := bq.db.Query(getNextUnfinishedBackfillState, true)
-		if err != nil || rows == nil {
-			bq.log.Error(err)
-			return
-		}
-		defer rows.Close()
-		if rows.Next() {
-			backfillState = bq.NewBackfillState(&PortalKey{}).Scan(rows)
-		}
-	}
-	return
+func (bq *BackfillQuery) GetNextUnfinishedBackfillState(ctx context.Context) (*BackfillState, error) {
+	return bq.QueryOne(ctx, getNextUnfinishedBackfillState)
+}
+
+type BackfillState struct {
+	qh *dbutil.QueryHelper[*BackfillState]
+
+	Portal            PortalKey
+	Dispatched        bool
+	BackfillComplete  bool
+	MessageCount      int
+	ImmediateComplete bool
+}
+
+func (b *BackfillState) Scan(row dbutil.Scannable) (*BackfillState, error) {
+	return dbutil.ValueOrErr(b, row.Scan(
+		&b.Portal.TeamID, &b.Portal.ChannelID, &b.Dispatched, &b.BackfillComplete, &b.MessageCount, &b.ImmediateComplete,
+	))
+}
+
+func (b *BackfillState) sqlVariables() []any {
+	return []any{b.Portal.TeamID, b.Portal.ChannelID, b.Dispatched, b.BackfillComplete, b.MessageCount, b.ImmediateComplete}
+}
+
+func (b *BackfillState) Upsert(ctx context.Context) error {
+	return b.qh.Exec(ctx, upsertBackfillStateQuery, b.sqlVariables()...)
+}
+
+func (b *BackfillState) SetDispatched(ctx context.Context, d bool) error {
+	b.Dispatched = d
+	return b.Upsert(ctx)
 }
