@@ -1,5 +1,5 @@
 // mautrix-slack - A Matrix-Slack puppeting bridge.
-// Copyright (C) 2022 Tulir Asokan
+// Copyright (C) 2024 Tulir Asokan
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as published by
@@ -17,110 +17,78 @@
 package database
 
 import (
-	"database/sql"
-	"errors"
+	"context"
 
-	log "maunium.net/go/maulogger/v2"
-
+	"go.mau.fi/util/dbutil"
 	"maunium.net/go/mautrix/id"
-	"maunium.net/go/mautrix/util/dbutil"
 )
 
+type ReactionQuery struct {
+	*dbutil.QueryHelper[*Reaction]
+}
+
+func newReaction(qh *dbutil.QueryHelper[*Reaction]) *Reaction {
+	return &Reaction{qh: qh}
+}
+
+const (
+	getReactionBaseQuery = `
+		SELECT team_id, channel_id, message_id, msg_first_part_id, author_id, emoji_id, mxid FROM reaction
+	`
+	getReactionBySlackIDQuery = getReactionBaseQuery + `
+		WHERE team_id=$1 AND channel_id=$2 AND message_id=$3 AND author_id=$4 AND emoji_id=$5
+	`
+	getReactionByMXIDQuery = getReactionBaseQuery + `
+		WHERE mxid=$1
+	`
+	insertReactionQuery = `
+		INSERT INTO reaction (team_id, channel_id, message_id, msg_first_part_id, author_id, emoji_id, mxid)
+		VALUES ($1, $2, $3, $4, $5, $6, $7)
+	`
+	deleteReactionQuery = `
+		DELETE FROM reaction WHERE team_id=$1 AND channel_id=$2 AND message_id=$3 AND author_id=$4 AND emoji_id=$5
+	`
+)
+
+func (rq *ReactionQuery) GetBySlackID(ctx context.Context, key PortalKey, messageID, authorID, emojiID string) (*Reaction, error) {
+	return rq.QueryOne(ctx, getReactionBySlackIDQuery, key.ChannelID, key.TeamID, messageID, authorID, emojiID)
+}
+
+func (rq *ReactionQuery) GetByMXID(ctx context.Context, eventID id.EventID) (*Reaction, error) {
+	return rq.QueryOne(ctx, getReactionByMXIDQuery, eventID)
+}
+
 type Reaction struct {
-	db  *Database
-	log log.Logger
+	qh *dbutil.QueryHelper[*Reaction]
 
-	Channel PortalKey
-
-	SlackMessageID string
-	MatrixEventID  id.EventID
-
-	// The slack ID of who create this reaction
-	AuthorID string
-
-	MatrixName string
-	MatrixURL  string // Used for custom emoji
-
-	SlackName string // The id or unicode of the emoji for slack
+	PortalKey
+	MessageID        string
+	MessageFirstPart PartID
+	AuthorID         string
+	EmojiID          string
+	MXID             id.EventID
 }
 
-func (r *Reaction) Scan(row dbutil.Scannable) *Reaction {
-	var slackName sql.NullString
-
-	err := row.Scan(
-		&r.Channel.TeamID, &r.Channel.ChannelID,
-		&r.SlackMessageID, &r.MatrixEventID,
+func (r *Reaction) Scan(row dbutil.Scannable) (*Reaction, error) {
+	return dbutil.ValueOrErr(r, row.Scan(
+		&r.TeamID,
+		&r.ChannelID,
+		&r.MessageID,
+		&r.MessageFirstPart,
 		&r.AuthorID,
-		&r.MatrixName, &r.MatrixURL,
-		&slackName)
-
-	if err != nil {
-		if !errors.Is(err, sql.ErrNoRows) {
-			r.log.Errorln("Database scan failed:", err)
-		}
-
-		return nil
-	}
-
-	r.SlackName = slackName.String
-
-	return r
+		&r.EmojiID,
+		&r.MXID,
+	))
 }
 
-func (r *Reaction) Insert(txn dbutil.Transaction) {
-	query := "INSERT INTO reaction" +
-		" (team_id, channel_id, slack_message_id, matrix_event_id," +
-		"  author_id, matrix_name, matrix_url, slack_name)" +
-		" VALUES($1, $2, $3, $4, $5, $6, $7, $8);"
-
-	var slackName sql.NullString
-
-	if r.SlackName != "" {
-		slackName = sql.NullString{String: r.SlackName, Valid: true}
-	}
-
-	args := []interface{}{r.Channel.TeamID, r.Channel.ChannelID,
-		r.SlackMessageID, r.MatrixEventID,
-		r.AuthorID,
-		r.MatrixName, r.MatrixURL,
-		slackName,
-	}
-
-	var err error
-	if txn != nil {
-		_, err = txn.Exec(query, args...)
-	} else {
-		_, err = r.db.Exec(query, args...)
-	}
-
-	if err != nil {
-		r.log.Warnfln("Failed to insert reaction for %s@%s: %v", r.Channel, r.SlackMessageID, err)
-	}
+func (r *Reaction) sqlVariables() []any {
+	return []any{r.TeamID, r.ChannelID, r.MessageID, r.MessageFirstPart, r.AuthorID, r.EmojiID, r.MXID}
 }
 
-func (r *Reaction) Update() {
-	// TODO: determine if we need this. The only scenario I can think of that
-	// would require this is if we insert a custom emoji before uploading to
-	// the homeserver?
+func (r *Reaction) Insert(ctx context.Context) error {
+	return r.qh.Exec(ctx, insertReactionQuery, r.sqlVariables()...)
 }
 
-func (r *Reaction) Delete() {
-	query := "DELETE FROM reaction WHERE" +
-		" team_id=$1 AND channel_id=$2 AND slack_message_id=$3 AND author_id=$4 AND slack_name=$5"
-
-	var slackName sql.NullString
-	if r.SlackName != "" {
-		slackName = sql.NullString{String: r.SlackName, Valid: true}
-	}
-
-	_, err := r.db.Exec(
-		query,
-		r.Channel.TeamID, r.Channel.ChannelID,
-		r.SlackMessageID, r.AuthorID,
-		slackName,
-	)
-
-	if err != nil {
-		r.log.Warnfln("Failed to delete reaction for %s@%s: %v", r.Channel, r.SlackMessageID, err)
-	}
+func (r *Reaction) Delete(ctx context.Context) error {
+	return r.qh.Exec(ctx, deleteReactionQuery, r.TeamID, r.ChannelID, r.MessageID, r.AuthorID, r.EmojiID)
 }
